Factor expand tests out of base
